refactor: add s01-s10 sequential prefixes to pipeline files
Number all pipeline files across src/data/, src/models/, and src/output/
to make execution order visible from filenames alone. Flatten
preprocessing/ subfolder into src/data/ since it contained only one file.
