+++ src/data/s02_loader.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import os
 from src.utils.weeks import DataOption
-from src.data.add_zero_weeks import AddWeeksWherePreapplicantsAreZero
+from src.data.s03_add_zero_weeks import AddWeeksWherePreapplicantsAreZero
 
 
 def get_path_latest(paths, data_option):
